# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Madhur Asha Enterprises Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /customers/{id}/push-zoho:
    post:
      operationId: pushCustomerToZoho
      tags: [customers]
      summary: Push customer to Zoho Books (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer synced to Zoho
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  zohoId:
                    type: string
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List items
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      operationId: createItem
      tags: [items]
      summary: Create an item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    put:
      operationId: updateItem
      tags: [items]
      summary: Update an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /items/{id}/push-zoho:
    post:
      operationId: pushItemToZoho
      tags: [items]
      summary: Push item to Zoho Books (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item synced to Zoho
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  zohoId:
                    type: string
  /items/bulk-push-zoho:
    post:
      operationId: bulkPushItemsToZoho
      tags: [items]
      summary: Bulk push items to Zoho Books (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                itemIds:
                  type: array
                  items:
                    type: integer
              required: [itemIds]
      responses:
        "200":
          description: Bulk push completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        itemId:
                          type: integer
                        itemName:
                          type: string
                        success:
                          type: boolean
                        zohoId:
                          type: string
                          nullable: true
                        error:
                          type: string
                          nullable: true
                  summary:
                    type: object
                    properties:
                      total:
                        type: integer
                      succeeded:
                        type: integer
                      failed:
                        type: integer
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          description: Business Name (Legal Name from GST)
        gstin:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
        contactPerson:
          type: string
          nullable: true
        gstStatus:
          type: string
          nullable: true
          description: GST registration status (Active/Cancelled)
        state:
          type: string
          nullable: true
        pincode:
          type: string
          nullable: true
        zohoId:
          type: string
          nullable: true
        zohoSyncStatus:
          type: string
          nullable: true
          enum: [pending, syncing, synced, error]
        zohoSyncedAt:
          type: string
          format: date-time
          nullable: true
        zohoErrorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        calculationCount:
          type: integer
      required: [id, name, createdAt, calculationCount]
    CreateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        gstin:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        contact:
          type: string
        contactPerson:
          type: string
          nullable: true
        gstStatus:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        pincode:
          type: string
          nullable: true
      required: [name]
...
    Item:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        hsnCode:
          type: string
          description: 4-8 digit HSN/SAC code
        description:
          type: string
          nullable: true
        unit:
          type: string
          enum: [Nos, Kg, Grams, Litre, Metre, Box, Bag, Piece, Set]
        purchaseRate:
          type: string
          description: Stored as string for precision
        sellingRate:
          type: string
          description: Stored as string for precision
        gstRate:
          type: string
          enum: ["0", "5", "12", "18", "28"]
        itemType:
          type: string
          enum: [goods, service]
        zohoId:
          type: string
          nullable: true
        zohoSyncStatus:
          type: string
          nullable: true
          enum: [pending, syncing, synced, error]
        zohoSyncedAt:
          type: string
          format: date-time
          nullable: true
        zohoErrorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, hsnCode, unit, purchaseRate, sellingRate, gstRate, itemType, createdAt, updatedAt]
    CreateItemRequest:
      type: object
      properties:
        name:
          type: string
        hsnCode:
          type: string
          pattern: '^\d{4,8}$'
          description: 4-8 digit HSN/SAC code
        description:
          type: string
          nullable: true
        unit:
          type: string
          enum: [Nos, Kg, Grams, Litre, Metre, Box, Bag, Piece, Set]
        purchaseRate:
          type: string
        sellingRate:
          type: string
        gstRate:
          type: string
          enum: ["0", "5", "12", "18", "28"]
        itemType:
          type: string
          enum: [goods, service]
      required: [name, hsnCode, unit, purchaseRate, sellingRate, gstRate, itemType]