# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Madhur Asha Enterprises Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemGstRate } from "./itemGstRate";
import type { ItemItemType } from "./itemItemType";
import type { ItemUnit } from "./itemUnit";
import type { ItemZohoSyncStatus } from "./itemZohoSyncStatus";

export interface Item {
  id: number;
  name: string;
  /** 4-8 digit HSN/SAC code */
  hsnCode: string;
  description?: string | null;
  unit: ItemUnit;
  /** Stored as string for precision */
  purchaseRate: string;
  /** Stored as string for precision */
  sellingRate: string;
  gstRate: ItemGstRate;
  itemType: ItemItemType;
  zohoId?: string | null;
  zohoSyncStatus?: ItemZohoSyncStatus;
  zohoSyncedAt?: Date | null;
  zohoErrorMessage?: string | null;
  createdAt: Date;
  updatedAt: Date;
}